用户测试

+++ src/main/java/com/test/testsystem/dto/KnowledegeQuestionCount.java
@@ -2,17 +2,19 @@ package com.test.testsystem.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class KnowledegeQuestionCount {
         private Integer knowledgeId;
         private String knowledgeName;
         private Integer questionId;
-        private Integer totalCount;
+        private BigInteger totalCount;
 
     public KnowledegeQuestionCount() {
     }
 
-    public KnowledegeQuestionCount(Integer knowledgeId, String knowledgeName, Integer questionId, Integer totalCount) {
+    public KnowledegeQuestionCount(Integer knowledgeId, String knowledgeName, Integer questionId, BigInteger totalCount) {
         this.knowledgeId = knowledgeId;
         this.knowledgeName = knowledgeName;
         this.questionId = questionId;
